+++ Changelog.md
@@ -1,5 +1,9 @@
 ## [2.1.0] - in progress
 * First release of Age of Mythology (AoM).
+* Python
+    * Generic function to display the BO design instructions.
+* AoE4
+    * Update images.
 
 ## [2.0.0] - 2024.08.21
 * New project folders structure.